refactor(shared-server): split group mutation reads and authority
