+++ packages/shared-server/rallar-system/group-state/mutation/idempotency/probe-group-mutation-idempotency.ts
@@ -1,0 +1,32 @@
+import { validateGroupMutationCommand } from '../command-validation/validate-group-mutation-command.ts';
+import type {
+    GroupMutationCommand,
+    GroupMutationIdempotencyProbe,
+    GroupMutationRead
+} from '../group-mutation-contracts.ts';
+import { validateCommandHash } from '../result-validation/validate-group-mutation-result.ts';
+import { validateGroupMutationRead } from '../state-validation/validate-group-mutation-read.ts';
+
+export function probeGroupMutationIdempotency(
+    command: GroupMutationCommand,
+    read: GroupMutationRead,
+    commandHash: string
+): GroupMutationIdempotencyProbe {
+    validateGroupMutationCommand(command);
+    validateGroupMutationRead(read, command);
+    validateCommandHash(commandHash, 'Group mutation commandHash');
+    if (!read.idempotency) {
+        return { outcome: 'miss' };
+    }
+    const record = read.idempotency.value;
+    if (record.receipt.commandId !== command.commandId) {
+        throw new TypeError('Stored group idempotency receipt command differs from command identity');
+    }
+    return record.commandHash === commandHash
+        ? { outcome: 'replay', receipt: record.receipt }
+        : {
+            outcome: 'idempotency-conflict',
+            existingCommandHash: record.commandHash,
+            receivedCommandHash: commandHash
+        };
+}
