+++ packages/shared-server/rallar-system/group-state/mutation/state-validation/validate-group-mutation.ts
@@ -2,6 +2,7 @@ import { jsonEquals } from '@shared/repository/state-utils.ts';
 
 import { requireJsonSafe } from '../../group-state-validation-primitives.ts';
 import { validatePresenceAdmission } from '../../persistence/validate-persisted-group-presence.ts';
+import { validateGroupMutationAuthority } from '../authority/validate-group-mutation-authority.ts';
 import { validateGroupMutationCommand } from '../command-validation/validate-group-mutation-command.ts';
 import type {
     GroupMutationCommand,
@@ -9,17 +10,15 @@ import type {
     GroupMutationFacts,
     GroupMutationRead
 } from '../group-mutation-contracts.ts';
-import {
-    computeGroupMutation,
-    validateFacts,
-    validateTrustedAuthorityMode
-} from '../orchestration/compute-group-mutation.ts';
+import { computeGroupMutation } from '../orchestration/compute-group-mutation.ts';
 import {
     validateComputedMutationShape,
     validateComputedOutboxEntries,
     validateComputedRosterFacts
 } from '../result-validation/validate-computed-group-mutation.ts';
+import { validateGroupMutationFacts } from './validate-group-mutation-facts.ts';
 import { validateGroupMutationRead } from './validate-group-mutation-read.ts';
+
 export function validateGroupMutation(
     input: Readonly<{
         command: GroupMutationCommand;
@@ -30,8 +29,8 @@ export function validateGroupMutation(
 ): void {
     validateGroupMutationCommand(input.command);
     validateGroupMutationRead(input.read, input.command);
-    validateFacts(input.facts);
-    validateTrustedAuthorityMode(input.command, input.facts);
+    validateGroupMutationFacts(input.facts);
+    validateGroupMutationAuthority(input.command, input.facts);
     requireJsonSafe(
         input.computed.outcome === 'write' ? { ...input.computed, outboxEntries: [] } : input.computed,
         'Group mutation computed result'
